Lots of changes and cleanups.

+++ java/tech/v2/datatype/IndexingSystem.java
@@ -5,7 +5,7 @@ public class IndexingSystem
 {
   public interface Backward
   {
-    //Returns an iterable that when asked returns an IntIter or nil
-    Object localToGlobal(int local_idx);
+    //Returns java Long or an Iterable
+    Object localToGlobal(long local_idx);
   }
 }
